backend/requirements.txt
# backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
psycopg2-binary==2.9.9
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
passlib==1.7.4
bcrypt==4.0.1